Fix: If the Hicks spot a spy at night, they don't enter combat as often.
The Hicks will still enter combat, as this is hardcoded - the fix merely causes them to do this only once instead of every turn.. Their behaviour of becoming hostile at night is hardcoded, inconsistent with other code behaviour, but should stay in. This poses no threat to a properly disguised spy and is merely a small annoyance.

git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13@7785 3b4a5df2-a311-0410-b5c6-a8a6f20db521